+++ frontend/interfaces/IPreviousPyconkr.ts
@@ -1,0 +1,5 @@
+import { PreviousPyconkr } from '../data/enums/PreviousPyconkr'
+
+export interface IPreviousPyconkr {
+    year: PreviousPyconkr
+}
