+++ frontend/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import type { NextPage } from 'next'
-import MainBackground from '../components/service/home/Main'
+import MainBackground from '../components/service/Home/Main'
 
 const Index: NextPage = () => {
     return (
